+++ git_machete/generated_docs.py
@@ -612,7 +612,7 @@ long_docs: Dict[str, str] = {
 
               `<PR-number-1> ... <PR-number-N>`    Pull request numbers to checkout.
 
-           `create-pr [--draft] [--title=<title>]`:
+           `create-pr [--draft] [--title=<title>] [--yes]`:
  
               Creates a PR for the current branch, using the upstream (parent) branch as the PR base.
               Once the PR is successfully created, annotates the current branch with the new PR's number.
@@ -633,6 +633,9 @@ long_docs: Dict[str, str] = {
                  <b>--title=<title></b>
 
               Set the PR title explicitly (the default is to use the first included commit's message as the title).
+
+                 <b>--yes</b>
+              Do not ask for confirmation whether to push the branch.
 
            `restack-pr`: