+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - added: `--removed-from-remote` flag to `git machete slide-out` (contributed by @raylu)
 - added: `--title` flag to `git machete github create-pr` that allows for setting PR title explicitly (suggested by @mjgigli)
 - added: `--with-urls` flag to `git machete github anno-prs` and `machete.github.annotateWithUrls` git config key that allow for adding the URL of the PR to the annotations (contributed by @guyboltonking)
+- added: `--yes` flag to `git machete github create-pr` so that the user isn't asked whether to push the branch (suggested by @mkondratek)
 - deprecated: `git machete clean` and `git machete github sync`; use `github checkout-prs --mine`, `delete-unmanaged` and `slide-out --removed-from-remote` instead
 - fixed: PR author is now always added to annotation if different from current user (contributed by @guyboltonking)